requirements.txt
backports.csv==1.0.7
boto3==1.10.22
botocore==1.13.22
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
configparser==4.0.2
cryptography==2.8
docutils==0.15.2
docx-mailmerge==0.5.0
Flask==1.1.1
Flask-HTTPAuth==3.3.0
future==0.18.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
pycparser==2.19
python-dateutil==2.8.0
python-dotenv==0.10.3
requests==2.22.0
s3transfer==0.2.1
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
